+++ som_autoreclama/tests/__init__.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import
 
+from .test_som_autoreclama_base import *
 from .test_som_autoreclama_states import *
+from .test_som_autoreclama_009 import *
