+++ som_autoreclama/tests/test_som_autoreclama_states.py
@@ -1,60 +1,9 @@
 # -*- coding: utf-8 -*-
-from destral import testing
-from destral.transaction import Transaction
-from datetime import date, timedelta, datetime
+from datetime import timedelta, datetime
 from addons import get_module_resource
 import mock
 from ..models import giscedata_atc, giscedata_polissa, som_autoreclama_state_history
-
-
-def today_str():
-    return date.today().strftime("%Y-%m-%d")
-
-
-def today_minus_str(d):
-    return (date.today() - timedelta(days=d)).strftime("%Y-%m-%d")
-
-
-class SomAutoreclamaBaseTests(testing.OOTestCase):
-    def setUp(self):
-        self.txn = Transaction().start(self.database)
-
-        self.cursor = self.txn.cursor
-        self.uid = self.txn.user
-
-    def tearDown(self):
-        self.txn.stop()
-
-    def get_model(self, model_name):
-        return self.openerp.pool.get(model_name)
-
-    def search_in(self, model, params):
-        model_obj = self.get_model(model)
-        found_ids = model_obj.search(self.cursor, self.uid, params)
-        return found_ids[0] if found_ids else None
-
-    def browse_referenced(self, reference):
-        model, id = reference.split(",")
-        model_obj = self.get_model(model)
-        return model_obj.browse(self.cursor, self.uid, int(id))
-
-    def get_object_reference(self, module, semantic_id):
-        ir_obj = self.get_model("ir.model.data")
-        expected_id = ir_obj.get_object_reference(
-            self.cursor, self.uid, module, semantic_id
-        )
-        return expected_id
-
-    def create_tags(self):
-        tag_obj = self.get_model("giscedata.atc.tag")
-        tag_obj.create(self.cursor, self.uid, {
-            'name': "[GET] Expedient FRAU",
-            'description': "Bla bla bla",
-        })
-        tag_obj.create(self.cursor, self.uid, {
-            'name': "[GET] Expedient ANOMALIA",
-            'description': "Bla bla bla",
-        })
+from test_som_autoreclama_base import SomAutoreclamaBaseTests, SomAutoreclamaEzATC_Test, today_str, today_minus_str  # noqa: E501
 
 
 class SomAutoreclamaStatesTest(SomAutoreclamaBaseTests):
@@ -1035,163 +984,6 @@ class SomAutoreclamaCreationWizardTest(SomAutoreclamaBaseTests):
         self.assertEqual(pas.subtipus_id.id, subtipus_id)
         self.assertEqual(pas.reclamacio_ids[0].num_factura, '9876543210ZYX')
         self.assertEqual(pas.comentaris, legal_txt)
-
-
-class SomAutoreclamaEzATC_Test(SomAutoreclamaBaseTests):
-    def build_atc(
-        self,
-        subtype="029",
-        r1=False,
-        channel="intercambi",
-        section="client",
-        log_days=3,
-        agent_actual="10",
-        state="pending",
-        active=True,
-        date_closed=None,
-        date=None,
-        polissa="polissa_0002",
-    ):
-        atc_obj = self.get_model("giscedata.atc")
-        _, polissa_id = self.get_object_reference(
-            "giscedata_polissa", polissa
-        )
-
-        par1_id = self.search_in("res.partner", [("name", "ilike", "Tiny sprl")])
-        par2_id = self.search_in("res.partner", [("name", "ilike", "ASUStek")])
-        par_obj = self.get_model("res.partner")
-        par_obj.write(self.cursor, self.uid, par1_id, {"ref": "58264"})
-        par_obj.write(self.cursor, self.uid, par2_id, {"ref": "58265"})
-
-        channel_id = self.search_in("res.partner.canal", [("name", "ilike", channel)])
-        section_id = self.search_in("crm.case.section", [("name", "ilike", section)])
-        subtipus_id = self.search_in("giscedata.subtipus.reclamacio", [("name", "=", subtype)])
-
-        new_case_data = {
-            "polissa_id": polissa_id,
-            "descripcio": u"Reclamació per retràs automàtica",
-            "canal_id": channel_id,
-            "section_id": section_id,
-            "subtipus_reclamacio_id": subtipus_id,
-            "comentaris": u"test test test",
-            "sense_responsable": True,
-            "tanca_al_finalitzar_r1": r1,
-            "crear_cas_r1": r1,
-        }
-        atc_id = atc_obj.create_general_atc_r1_case_via_wizard(
-            self.cursor, self.uid, new_case_data, {}
-        )
-        last_write = {
-            "agent_actual": agent_actual,
-            "state": state,
-            "active": active,
-        }
-        if date:
-            last_write['date'] = date
-        atc_obj.write(
-            self.cursor,
-            self.uid,
-            atc_id,
-            last_write,
-        )
-        if date_closed:
-            atc_obj.write(
-                self.cursor,
-                self.uid,
-                atc_id,
-                {'date_closed': date_closed}
-            )
-        atc = atc_obj.browse(self.cursor, self.uid, atc_id)
-        log_obj = self.get_model("crm.case.log")
-        log_obj.write(self.cursor, self.uid, atc.log_ids[1].id, {"date": today_minus_str(log_days)})
-
-        return atc_id
-
-    def build_polissa(
-        self,
-        name="polissa_0002",
-        f1_date_days_from_today=None,
-        initial_state=None,
-        data_baixa=None,
-        data_baixa_from_today=None,
-    ):
-        f1i_obj = self.get_model("giscedata.polissa.f1.info")
-        h_obj = self.get_model("som.autoreclama.state.history.polissa")
-
-        _, polissa_id = self.get_object_reference(
-            "giscedata_polissa", name
-        )
-
-        if f1_date_days_from_today is not None:
-            date = today_minus_str(f1_date_days_from_today)
-            f1i_obj.create(
-                self.cursor,
-                self.uid,
-                {
-                    'polissa_id': polissa_id,
-                    'data_ultima_lectura_f1': date,
-                }
-            )
-
-        states = {
-            'correct': "correct_state_workflow_polissa",
-            'loop': "loop_state_workflow_polissa",
-            'disabled': "disabled_state_workflow_polissa",
-        }
-        if initial_state in states:
-            _, st_id = self.get_object_reference(
-                "som_autoreclama",
-                states[initial_state]
-            )
-            h_obj.historize(
-                self.cursor, self.uid,
-                polissa_id,
-                st_id,
-                None,
-                None
-            )
-
-        par1_id = self.search_in("res.partner", [("name", "ilike", "Tiny sprl")])
-        par2_id = self.search_in("res.partner", [("name", "ilike", "ASUStek")])
-        par_obj = self.get_model("res.partner")
-        par_obj.write(self.cursor, self.uid, par1_id, {"ref": "58264"})
-        par_obj.write(self.cursor, self.uid, par2_id, {"ref": "58265"})
-
-        pol_obj = self.get_model("giscedata.polissa")
-        pol_obj.write(self.cursor, self.uid, polissa_id, {"state": 'activa'})
-        if data_baixa_from_today:
-            data_baixa = today_minus_str(data_baixa_from_today)
-
-        if data_baixa is not None:
-            vals = {"data_baixa": data_baixa}
-            if data_baixa:
-                vals['state'] = 'baixa'
-
-            pol_obj.write(self.cursor, self.uid, polissa_id, vals)
-
-        return polissa_id
-
-    def add_done_006_to_polissa(self, pol_id, days):
-        dat = today_minus_str(days)
-        return self.build_atc(subtype="006", state='done', date_closed=dat, date=dat, r1=False)
-
-    def add_correct_to_history(self, pol_id, days, atc_id=None):
-        imd_obj = self.get_model("ir.model.data")
-        polh_obj = self.get_model("som.autoreclama.state.history.polissa")
-        correct_state_id = imd_obj.get_object_reference(
-            self.cursor, self.uid, "som_autoreclama", "correct_state_workflow_polissa"
-        )[1]
-        polh_obj.historize(
-            self.cursor, self.uid, pol_id, correct_state_id, today_minus_str(days), atc_id)
-
-    def add_review_to_history(self, pol_id, days, atc_id=None):
-        imd_obj = self.get_model("ir.model.data")
-        polh_obj = self.get_model("som.autoreclama.state.history.polissa")
-        review_state_id = imd_obj.get_object_reference(
-            self.cursor, self.uid, "som_autoreclama", "review_state_workflow_polissa"
-        )[1]
-        polh_obj.historize(
-            self.cursor, self.uid, pol_id, review_state_id, today_minus_str(days), atc_id)
 
 
 class SomAutoreclamaConditionsTest(SomAutoreclamaEzATC_Test):
@@ -2648,833 +2440,3 @@ class SomAutoreclamaf1cAutomationTest(SomAutoreclamaEzATC_Test):
         self.assertTrue(exce in msg)
         exce = u"Error en la creació del CAC amb R1 010, F1 no es tipus anomalia o frau id_f1"
         self.assertTrue(exce in msg)
-
-
-class SomAutoreclama009AutomationTest(SomAutoreclamaEzATC_Test):
-
-    def test_create_ATC_R1_009_from_polissa_via_wizard__bad_f1(self):
-        atc_obj = self.get_model("giscedata.atc")
-        pol_obj = self.get_model("giscedata.polissa")
-        f1_obj = self.get_model("giscedata.facturacio.importacio.linia")
-
-        _, polissa_id = self.get_object_reference(
-            "giscedata_polissa", "polissa_0004"
-        )
-        pol = pol_obj.browse(self.cursor, self.uid, polissa_id)
-        f1_ids = f1_obj.search(self.cursor, self.uid, [("cups_id", "in", [pol.cups.id])])
-        self.assertGreater(len(f1_ids), 0)
-        f1_id = f1_ids[0]
-
-        f1 = f1_obj.browse(self.cursor, self.uid, f1_id)
-        self.assertEqual(f1.cups_id.name, pol.cups.name)
-
-        f1_obj.write(self.cursor, self.uid, f1_id, {
-            "invoice_number_text": "1234567890ABCD",
-            "type_factura": "C",
-        })
-
-        with self.assertRaises(Exception) as context:
-            _ = atc_obj.create_ATC_R1_009_from_polissa_via_wizard(
-                self.cursor, self.uid, polissa_id, {}
-            )
-
-        msg = u"Error en la creació del CAC R1 009, no s'ha trobat F1 adient!!!"
-        self.assertTrue(msg in context.exception.message)
-
-    def test_create_ATC_R1_009_from_polissa_via_wizard__no_provider_invoice(self):
-        atc_obj = self.get_model("giscedata.atc")
-        pol_obj = self.get_model("giscedata.polissa")
-        f1_obj = self.get_model("giscedata.facturacio.importacio.linia")
-
-        _, polissa_id = self.get_object_reference(
-            "giscedata_polissa", "polissa_0004"
-        )
-        pol = pol_obj.browse(self.cursor, self.uid, polissa_id)
-        f1_ids = f1_obj.search(self.cursor, self.uid, [("cups_id", "in", [pol.cups.id])])
-        self.assertGreater(len(f1_ids), 0)
-        f1_id = f1_ids[0]
-
-        f1 = f1_obj.browse(self.cursor, self.uid, f1_id)
-        self.assertEqual(f1.cups_id.name, pol.cups.name)
-
-        f1_obj.write(self.cursor, self.uid, f1_id, {
-            "invoice_number_text": "1234567890ABCD",
-            "type_factura": "N",
-            "polissa_id": pol.id,
-            "fecha_factura_desde": "2015-01-01",
-            "fecha_factura_hasta": "2015-01-31",
-            "import_phase": 50,
-        })
-
-        with self.assertRaises(Exception) as context:
-            _ = atc_obj.create_ATC_R1_009_from_polissa_via_wizard(
-                self.cursor, self.uid, polissa_id, {}
-            )
-
-        msg = u"Error en la creació del CAC R1 009, l'F1 trobat no té factura de proveïdor id f1"
-        self.assertTrue(msg in context.exception.message)
-
-    def test_create_ATC_R1_009_from_polissa_via_wizard__one_ok(self):
-        atc_obj = self.get_model("giscedata.atc")
-        pol_obj = self.get_model("giscedata.polissa")
-        f1_obj = self.get_model("giscedata.facturacio.importacio.linia")
-        ff_obj = self.get_model("giscedata.facturacio.importacio.linia.factura")
-
-        _, polissa_id = self.get_object_reference(
-            "giscedata_polissa", "polissa_0004"
-        )
-        pol = pol_obj.browse(self.cursor, self.uid, polissa_id)
-        f1_ids = f1_obj.search(self.cursor, self.uid, [("cups_id", "in", [pol.cups.id])])
-        self.assertGreater(len(f1_ids), 0)
-        f1_id = f1_ids[0]
-
-        f1 = f1_obj.browse(self.cursor, self.uid, f1_id)
-        self.assertEqual(f1.cups_id.name, pol.cups.name)
-
-        f1_obj.write(self.cursor, self.uid, f1_id, {
-            "invoice_number_text": "1234567890ABCD",
-            "type_factura": "N",
-            "polissa_id": pol.id,
-            "fecha_factura_desde": "2015-01-01",
-            "fecha_factura_hasta": "2015-01-31",
-            "import_phase": 50,
-        })
-        ff_obj.create(self.cursor, self.uid, {
-            'linia_id': f1_id,
-            'tipo_factura': '04',
-            'importacio_id': 1,
-            'address_invoice_id': pol.direccio_pagament.id,
-            'partner_id': pol.titular.id,
-            'account_id': 3,
-            'polissa_id': polissa_id,
-            'tarifa_acces_id': pol.tarifa.id,
-            'cups_id': pol.cups.id,
-            'factura_id': 3,
-            'llista_preu': pol.llista_preu.id,
-        })
-
-        atc_id = atc_obj.create_ATC_R1_009_from_polissa_via_wizard(
-            self.cursor, self.uid, polissa_id, {}
-        )
-
-        atc = atc_obj.browse(self.cursor, self.uid, atc_id)
-
-        channel_id = self.search_in("res.partner.canal", [("name", "ilike", "intercambi")])
-        subtipus_id = self.search_in("giscedata.subtipus.reclamacio", [("name", "=", "009")])
-        _, section_id = self.get_object_reference(
-            "som_switching", "atc_section_factura"
-        )
-        tag_id = self.search_in("giscedata.atc.tag", [("name", "ilike", "%Autocac 009")])
-
-        self.assertEqual(atc.name, u"AUTOCAC 009")
-        self.assertEqual(atc.canal_id.id, channel_id)
-        self.assertEqual(atc.section_id.id, section_id)
-        self.assertEqual(atc.subtipus_id.id, subtipus_id)
-        self.assertEqual(atc.polissa_id.id, polissa_id)
-        self.assertEqual(atc.tag.id, tag_id)
-
-        self.assertEqual(atc.tancar_cac_al_finalitzar_r1, True)
-        self.assertEqual(atc.state, "pending")
-        self.assertEqual(atc.agent_actual, "10")
-
-        model, id = atc.ref.split(",")
-        self.assertEqual(model, "giscedata.switching")
-        model_obj = self.get_model(model)
-        ref = model_obj.browse(self.cursor, self.uid, int(id))
-
-        self.assertEqual(ref.proces_id.name, u"R1")
-        self.assertEqual(ref.step_id.name, u"01")
-        self.assertEqual(ref.section_id.name, u"Switching")
-        self.assertEqual(ref.cups_polissa_id.id, polissa_id)
-        self.assertEqual(ref.state, u"open")
-        self.assertEqual(ref.ref, u"giscedata.atc, {}".format(atc.id))
-
-        model, id = ref.step_ids[0].pas_id.split(",")
-        self.assertEqual(model, "giscedata.switching.r1.01")
-        model_obj = self.get_model(model)
-        pas = model_obj.browse(self.cursor, self.uid, int(id))
-
-        self.assertEqual(pas.subtipus_id.id, subtipus_id)
-        self.assertEqual(pas.reclamacio_ids[0].num_factura, '1234567890ABCD')
-
-    def test_create_ATC_R1_009_from_polissa_via_wizard__use_last_valid_f1(self):
-        atc_obj = self.get_model("giscedata.atc")
-        pol_obj = self.get_model("giscedata.polissa")
-        f1_obj = self.get_model("giscedata.facturacio.importacio.linia")
-        ff_obj = self.get_model("giscedata.facturacio.importacio.linia.factura")
-
-        _, polissa_id = self.get_object_reference(
-            "giscedata_polissa", "polissa_0004"
-        )
-        pol = pol_obj.browse(self.cursor, self.uid, polissa_id)
-        f1_ids = f1_obj.search(self.cursor, self.uid, [("cups_id", "in", [pol.cups.id])])
-        self.assertGreater(len(f1_ids), 1)
-        f1_id = f1_ids[0]
-
-        f1 = f1_obj.browse(self.cursor, self.uid, f1_id)
-        self.assertEqual(f1.cups_id.name, pol.cups.name)
-
-        f1_obj.write(self.cursor, self.uid, f1_id, {
-            "invoice_number_text": "1234567890ABCD",
-            "type_factura": "N",
-            "polissa_id": pol.id,
-            "fecha_factura_desde": "2015-01-01",
-            "fecha_factura_hasta": "2015-01-31",
-            "import_phase": 50,
-        })
-        ff_obj.create(self.cursor, self.uid, {
-            'linia_id': f1_id,
-            'tipo_factura': '04',
-            'importacio_id': 1,
-            'address_invoice_id': pol.direccio_pagament.id,
-            'partner_id': pol.titular.id,
-            'account_id': 3,
-            'polissa_id': polissa_id,
-            'tarifa_acces_id': pol.tarifa.id,
-            'cups_id': pol.cups.id,
-            'factura_id': 3,
-            'llista_preu': pol.llista_preu.id,
-        })
-
-        f1_id = f1_ids[1]
-
-        f1 = f1_obj.browse(self.cursor, self.uid, f1_id)
-        self.assertEqual(f1.cups_id.name, pol.cups.name)
-
-        f1_obj.write(self.cursor, self.uid, f1_id, {
-            "invoice_number_text": "ABCD987654321",
-            "type_factura": "N",
-            "polissa_id": pol.id,
-            "fecha_factura_desde": "2015-02-01",
-            "fecha_factura_hasta": "2015-02-28",
-            "import_phase": 40,
-        })
-        ff_obj.create(self.cursor, self.uid, {
-            'linia_id': f1_id,
-            'tipo_factura': '04',
-            'importacio_id': 1,
-            'address_invoice_id': pol.direccio_pagament.id,
-            'partner_id': pol.titular.id,
-            'account_id': 3,
-            'polissa_id': polissa_id,
-            'tarifa_acces_id': pol.tarifa.id,
-            'cups_id': pol.cups.id,
-            'factura_id': 3,
-            'llista_preu': pol.llista_preu.id,
-        })
-
-        atc_id = atc_obj.create_ATC_R1_009_from_polissa_via_wizard(
-            self.cursor, self.uid, polissa_id, {}
-        )
-
-        atc = atc_obj.browse(self.cursor, self.uid, atc_id)
-
-        channel_id = self.search_in("res.partner.canal", [("name", "ilike", "intercambi")])
-        subtipus_id = self.search_in("giscedata.subtipus.reclamacio", [("name", "=", "009")])
-        _, section_id = self.get_object_reference(
-            "som_switching", "atc_section_factura"
-        )
-        tag_id = self.search_in("giscedata.atc.tag", [("name", "ilike", "%Autocac 009")])
-
-        self.assertEqual(atc.name, u"AUTOCAC 009")
-        self.assertEqual(atc.canal_id.id, channel_id)
-        self.assertEqual(atc.section_id.id, section_id)
-        self.assertEqual(atc.subtipus_id.id, subtipus_id)
-        self.assertEqual(atc.polissa_id.id, polissa_id)
-        self.assertEqual(atc.tag.id, tag_id)
-
-        self.assertEqual(atc.tancar_cac_al_finalitzar_r1, True)
-        self.assertEqual(atc.state, "pending")
-        self.assertEqual(atc.agent_actual, "10")
-
-        model, id = atc.ref.split(",")
-        self.assertEqual(model, "giscedata.switching")
-        model_obj = self.get_model(model)
-        ref = model_obj.browse(self.cursor, self.uid, int(id))
-
-        self.assertEqual(ref.proces_id.name, u"R1")
-        self.assertEqual(ref.step_id.name, u"01")
-        self.assertEqual(ref.section_id.name, u"Switching")
-        self.assertEqual(ref.cups_polissa_id.id, polissa_id)
-        self.assertEqual(ref.state, u"open")
-        self.assertEqual(ref.ref, u"giscedata.atc, {}".format(atc.id))
-
-        model, id = ref.step_ids[0].pas_id.split(",")
-        self.assertEqual(model, "giscedata.switching.r1.01")
-        model_obj = self.get_model(model)
-        pas = model_obj.browse(self.cursor, self.uid, int(id))
-
-        self.assertEqual(pas.subtipus_id.id, subtipus_id)
-        self.assertEqual(pas.reclamacio_ids[0].num_factura, 'ABCD987654321')
-
-    def test_create_ATC_R1_009_from_polissa_via_wizard__use_last_valid_f1_complex(self):
-        atc_obj = self.get_model("giscedata.atc")
-        pol_obj = self.get_model("giscedata.polissa")
-        f1_obj = self.get_model("giscedata.facturacio.importacio.linia")
-        ff_obj = self.get_model("giscedata.facturacio.importacio.linia.factura")
-
-        _, polissa_id = self.get_object_reference(
-            "giscedata_polissa", "polissa_0004"
-        )
-        pol = pol_obj.browse(self.cursor, self.uid, polissa_id)
-        f1_ids = f1_obj.search(self.cursor, self.uid, [("cups_id", "in", [pol.cups.id])])
-        self.assertGreater(len(f1_ids), 2)
-        f1_id = f1_ids[0]
-
-        f1 = f1_obj.browse(self.cursor, self.uid, f1_id)
-        self.assertEqual(f1.cups_id.name, pol.cups.name)
-
-        f1_obj.write(self.cursor, self.uid, f1_id, {
-            "invoice_number_text": "1234567890ABCD",
-            "type_factura": "R",
-            "polissa_id": pol.id,
-            "fecha_factura_desde": "2015-01-01",
-            "fecha_factura_hasta": "2015-01-31",
-            "import_phase": 50,
-        })
-        ff_obj.create(self.cursor, self.uid, {
-            'linia_id': f1_id,
-            'tipo_factura': '04',
-            'importacio_id': 1,
-            'address_invoice_id': pol.direccio_pagament.id,
-            'partner_id': pol.titular.id,
-            'account_id': 3,
-            'polissa_id': polissa_id,
-            'tarifa_acces_id': pol.tarifa.id,
-            'cups_id': pol.cups.id,
-            'factura_id': 3,
-            'llista_preu': pol.llista_preu.id,
-        })
-
-        f1_id = f1_ids[1]
-
-        f1 = f1_obj.browse(self.cursor, self.uid, f1_id)
-        self.assertEqual(f1.cups_id.name, pol.cups.name)
-
-        f1_obj.write(self.cursor, self.uid, f1_id, {
-            "invoice_number_text": "ABCD987654321",
-            "type_factura": "G",
-            "polissa_id": pol.id,
-            "fecha_factura_desde": "2015-02-01",
-            "fecha_factura_hasta": "2015-02-28",
-            "import_phase": 30,
-        })
-        ff_obj.create(self.cursor, self.uid, {
-            'linia_id': f1_id,
-            'tipo_factura': '04',
-            'importacio_id': 1,
-            'address_invoice_id': pol.direccio_pagament.id,
-            'partner_id': pol.titular.id,
-            'account_id': 3,
-            'polissa_id': polissa_id,
-            'tarifa_acces_id': pol.tarifa.id,
-            'cups_id': pol.cups.id,
-            'factura_id': 3,
-            'llista_preu': pol.llista_preu.id,
-        })
-
-        f1_id = f1_ids[2]
-
-        f1 = f1_obj.browse(self.cursor, self.uid, f1_id)
-        self.assertEqual(f1.cups_id.name, pol.cups.name)
-
-        f1_obj.write(self.cursor, self.uid, f1_id, {
-            "invoice_number_text": "FFFJJJJJJJJJJ",
-            "type_factura": "N",
-            "polissa_id": pol.id,
-            "fecha_factura_desde": "2015-03-01",
-            "fecha_factura_hasta": "2015-03-30",
-            "import_phase": 20,
-        })
-        ff_obj.create(self.cursor, self.uid, {
-            'linia_id': f1_id,
-            'tipo_factura': '04',
-            'importacio_id': 1,
-            'address_invoice_id': pol.direccio_pagament.id,
-            'partner_id': pol.titular.id,
-            'account_id': 3,
-            'polissa_id': polissa_id,
-            'tarifa_acces_id': pol.tarifa.id,
-            'cups_id': pol.cups.id,
-            'factura_id': 3,
-            'llista_preu': pol.llista_preu.id,
-        })
-
-        atc_id = atc_obj.create_ATC_R1_009_from_polissa_via_wizard(
-            self.cursor, self.uid, polissa_id, {}
-        )
-
-        atc = atc_obj.browse(self.cursor, self.uid, atc_id)
-
-        channel_id = self.search_in("res.partner.canal", [("name", "ilike", "intercambi")])
-        subtipus_id = self.search_in("giscedata.subtipus.reclamacio", [("name", "=", "009")])
-        _, section_id = self.get_object_reference(
-            "som_switching", "atc_section_factura"
-        )
-
-        self.assertEqual(atc.name, u"AUTOCAC 009")
-        self.assertEqual(atc.canal_id.id, channel_id)
-        self.assertEqual(atc.section_id.id, section_id)
-        self.assertEqual(atc.subtipus_id.id, subtipus_id)
-        self.assertEqual(atc.polissa_id.id, polissa_id)
-        self.assertTrue(u"Autocac 009" in atc.tag.name)
-
-        self.assertEqual(atc.tancar_cac_al_finalitzar_r1, True)
-        self.assertEqual(atc.state, "pending")
-        self.assertEqual(atc.agent_actual, "10")
-
-        model, id = atc.ref.split(",")
-        self.assertEqual(model, "giscedata.switching")
-        model_obj = self.get_model(model)
-        ref = model_obj.browse(self.cursor, self.uid, int(id))
-
-        self.assertEqual(ref.proces_id.name, u"R1")
-        self.assertEqual(ref.step_id.name, u"01")
-        self.assertEqual(ref.section_id.name, u"Switching")
-        self.assertEqual(ref.cups_polissa_id.id, polissa_id)
-        self.assertEqual(ref.state, u"open")
-        self.assertEqual(ref.ref, u"giscedata.atc, {}".format(atc.id))
-
-        model, id = ref.step_ids[0].pas_id.split(",")
-        self.assertEqual(model, "giscedata.switching.r1.01")
-        model_obj = self.get_model(model)
-        pas = model_obj.browse(self.cursor, self.uid, int(id))
-
-        self.assertEqual(pas.subtipus_id.id, subtipus_id)
-        self.assertEqual(pas.reclamacio_ids[0].num_factura, 'ABCD987654321')
-
-    def test_create_ATC_R1_009_from_polissa_via_wizard__error_has_a_previous_009(self):
-        atc_obj = self.get_model("giscedata.atc")
-        pol_obj = self.get_model("giscedata.polissa")
-        f1_obj = self.get_model("giscedata.facturacio.importacio.linia")
-        ff_obj = self.get_model("giscedata.facturacio.importacio.linia.factura")
-
-        _, polissa_id = self.get_object_reference(
-            "giscedata_polissa", "polissa_0004"
-        )
-        pol = pol_obj.browse(self.cursor, self.uid, polissa_id)
-        f1_ids = f1_obj.search(self.cursor, self.uid, [("cups_id", "in", [pol.cups.id])])
-        self.assertGreater(len(f1_ids), 0)
-        f1_id = f1_ids[0]
-
-        f1 = f1_obj.browse(self.cursor, self.uid, f1_id)
-        self.assertEqual(f1.cups_id.name, pol.cups.name)
-
-        f1_obj.write(self.cursor, self.uid, f1_id, {
-            "invoice_number_text": "1234567890ABCD",
-            "type_factura": "R",
-            "polissa_id": pol.id,
-            "fecha_factura_desde": "2015-01-01",
-            "fecha_factura_hasta": "2015-01-31",
-            "import_phase": 50,
-        })
-        ff_obj.create(self.cursor, self.uid, {
-            'linia_id': f1_id,
-            'tipo_factura': '04',
-            'importacio_id': 1,
-            'address_invoice_id': pol.direccio_pagament.id,
-            'partner_id': pol.titular.id,
-            'account_id': 3,
-            'polissa_id': polissa_id,
-            'tarifa_acces_id': pol.tarifa.id,
-            'cups_id': pol.cups.id,
-            'factura_id': 3,
-            'llista_preu': pol.llista_preu.id,
-        })
-
-        atc_id = atc_obj.create_ATC_R1_009_from_polissa_via_wizard(
-            self.cursor, self.uid, polissa_id, {}
-        )
-
-        with self.assertRaises(Exception) as context:
-            _ = atc_obj.create_ATC_R1_009_from_polissa_via_wizard(
-                self.cursor, self.uid, polissa_id, {}
-            )
-
-        msg = u"Error en la creació del CAC R1 009, ja n'hi ha un CAC  009 en estat obert o pendent amb id {}!!!".format(atc_id)  # noqa: E501
-        self.assertTrue(msg in context.exception.message)
-
-    def test_create_ATC_R1_009_from_polissa_via_wizard__previous_009(self):
-        atc_obj = self.get_model("giscedata.atc")
-        pol_obj = self.get_model("giscedata.polissa")
-        f1_obj = self.get_model("giscedata.facturacio.importacio.linia")
-        ff_obj = self.get_model("giscedata.facturacio.importacio.linia.factura")
-
-        _, polissa_id = self.get_object_reference(
-            "giscedata_polissa", "polissa_0004"
-        )
-        pol = pol_obj.browse(self.cursor, self.uid, polissa_id)
-        f1_ids = f1_obj.search(self.cursor, self.uid, [("cups_id", "in", [pol.cups.id])])
-        self.assertGreater(len(f1_ids), 0)
-        f1_id = f1_ids[0]
-
-        f1 = f1_obj.browse(self.cursor, self.uid, f1_id)
-        self.assertEqual(f1.cups_id.name, pol.cups.name)
-
-        f1_obj.write(self.cursor, self.uid, f1_id, {
-            "invoice_number_text": "1234567890ABCD",
-            "type_factura": "R",
-            "polissa_id": pol.id,
-            "fecha_factura_desde": "2015-01-01",
-            "fecha_factura_hasta": "2015-01-31",
-            "import_phase": 50,
-        })
-        ff_obj.create(self.cursor, self.uid, {
-            'linia_id': f1_id,
-            'tipo_factura': '04',
-            'importacio_id': 1,
-            'address_invoice_id': pol.direccio_pagament.id,
-            'partner_id': pol.titular.id,
-            'account_id': 3,
-            'polissa_id': polissa_id,
-            'tarifa_acces_id': pol.tarifa.id,
-            'cups_id': pol.cups.id,
-            'factura_id': 3,
-            'llista_preu': pol.llista_preu.id,
-        })
-
-        atc_id = self.build_atc(log_days=60, subtype="009", polissa="polissa_0004")
-
-        with self.assertRaises(Exception) as context:
-            _ = atc_obj.create_ATC_R1_009_from_polissa_via_wizard(
-                self.cursor, self.uid, polissa_id, {}
-            )
-
-        msg = u"Error en la creació del CAC R1 009, ja n'hi ha un CAC  009 en estat obert o pendent amb id {}!!!".format(atc_id)  # noqa: E501
-        self.assertTrue(msg in context.exception.message)
-
-    def test_create_ATC_R1_009_from_polissa_via_wizard__previous_036(self):
-        atc_obj = self.get_model("giscedata.atc")
-        pol_obj = self.get_model("giscedata.polissa")
-        f1_obj = self.get_model("giscedata.facturacio.importacio.linia")
-        ff_obj = self.get_model("giscedata.facturacio.importacio.linia.factura")
-
-        _, polissa_id = self.get_object_reference(
-            "giscedata_polissa", "polissa_0004"
-        )
-        pol = pol_obj.browse(self.cursor, self.uid, polissa_id)
-        f1_ids = f1_obj.search(self.cursor, self.uid, [("cups_id", "in", [pol.cups.id])])
-        self.assertGreater(len(f1_ids), 0)
-        f1_id = f1_ids[0]
-
-        f1 = f1_obj.browse(self.cursor, self.uid, f1_id)
-        self.assertEqual(f1.cups_id.name, pol.cups.name)
-
-        f1_obj.write(self.cursor, self.uid, f1_id, {
-            "invoice_number_text": "1234567890ABCD",
-            "type_factura": "R",
-            "polissa_id": pol.id,
-            "fecha_factura_desde": "2015-01-01",
-            "fecha_factura_hasta": "2015-01-31",
-            "import_phase": 50,
-        })
-        ff_obj.create(self.cursor, self.uid, {
-            'linia_id': f1_id,
-            'tipo_factura': '04',
-            'importacio_id': 1,
-            'address_invoice_id': pol.direccio_pagament.id,
-            'partner_id': pol.titular.id,
-            'account_id': 3,
-            'polissa_id': polissa_id,
-            'tarifa_acces_id': pol.tarifa.id,
-            'cups_id': pol.cups.id,
-            'factura_id': 3,
-            'llista_preu': pol.llista_preu.id,
-        })
-
-        atc_id = self.build_atc(log_days=60, subtype="036", polissa="polissa_0004")
-
-        with self.assertRaises(Exception) as context:
-            _ = atc_obj.create_ATC_R1_009_from_polissa_via_wizard(
-                self.cursor, self.uid, polissa_id, {}
-            )
-
-        msg = u"Error en la creació del CAC R1 009, ja n'hi ha un CAC 036 en estat obert o pendent amb id {}!!!".format(atc_id)  # noqa: E501
-        self.assertTrue(msg in context.exception.message)
-
-
-"""
-    def test_create_ATC_R1_009_from_polissa_via_wizard__previous_036_closed(self):
-        atc_obj = self.get_model("giscedata.atc")
-        pol_obj = self.get_model("giscedata.polissa")
-        f1_obj = self.get_model("giscedata.facturacio.importacio.linia")
-        ff_obj = self.get_model("giscedata.facturacio.importacio.linia.factura")
-
-        _, polissa_id = self.get_object_reference(
-            "giscedata_polissa", "polissa_0004"
-        )
-        pol = pol_obj.browse(self.cursor, self.uid, polissa_id)
-        f1_ids = f1_obj.search(self.cursor, self.uid, [("cups_id", "in", [pol.cups.id])])
-        self.assertGreater(len(f1_ids), 0)
-        f1_id = f1_ids[0]
-
-        f1 = f1_obj.browse(self.cursor, self.uid, f1_id)
-        self.assertEqual(f1.cups_id.name, pol.cups.name)
-
-        f1_obj.write(self.cursor, self.uid, f1_id, {
-            "invoice_number_text": "1234567890ABCD",
-            "type_factura": "R",
-            "polissa_id": pol.id,
-            "fecha_factura_desde": "2015-01-01",
-            "fecha_factura_hasta": "2015-01-31",
-            "import_phase": 50,
-        })
-        ff_obj.create(self.cursor, self.uid, {
-            'linia_id': f1_id,
-            'tipo_factura': '04',
-            'importacio_id': 1,
-            'address_invoice_id': pol.direccio_pagament.id,
-            'partner_id': pol.titular.id,
-            'account_id': 3,
-            'polissa_id': polissa_id,
-            'tarifa_acces_id': pol.tarifa.id,
-            'cups_id': pol.cups.id,
-            'factura_id': 3,
-            'llista_preu': pol.llista_preu.id,
-        })
-
-        _ = self.build_atc(log_days=60, subtype="036", polissa="polissa_0004",
-                           state='done', date_closed='2014-11-10')
-
-        atc_id = atc_obj.create_ATC_R1_009_from_polissa_via_wizard(
-            self.cursor, self.uid, polissa_id, {}
-        )
-
-        atc = atc_obj.browse(self.cursor, self.uid, atc_id)
-        self.assertEqual(atc.name, u"AUTOCAC 009")
-
-    def test_create_ATC_R1_009_from_polissa_via_wizard__tag_one_indexed(self):
-        atc_obj = self.get_model("giscedata.atc")
-        pol_obj = self.get_model("giscedata.polissa")
-        f1_obj = self.get_model("giscedata.facturacio.importacio.linia")
-        ff_obj = self.get_model("giscedata.facturacio.importacio.linia.factura")
-
-        _, polissa_id = self.get_object_reference(
-            "giscedata_polissa", "polissa_0004"
-        )
-        pol_obj.write(self.cursor, self.uid, polissa_id, {'mode_facturacio': 'index'})
-
-        pol = pol_obj.browse(self.cursor, self.uid, polissa_id)
-        f1_ids = f1_obj.search(self.cursor, self.uid, [("cups_id", "in", [pol.cups.id])])
-        self.assertGreater(len(f1_ids), 0)
-        f1_id = f1_ids[0]
-
-        f1 = f1_obj.browse(self.cursor, self.uid, f1_id)
-        self.assertEqual(f1.cups_id.name, pol.cups.name)
-
-        f1_obj.write(self.cursor, self.uid, f1_id, {
-            "invoice_number_text": "1234567890ABCD",
-            "type_factura": "R",
-            "polissa_id": pol.id,
-            "fecha_factura_desde": "2015-01-01",
-            "fecha_factura_hasta": "2015-01-31",
-            "import_phase": 50,
-        })
-        ff_obj.create(self.cursor, self.uid, {
-            'linia_id': f1_id,
-            'tipo_factura': '04',
-            'importacio_id': 1,
-            'address_invoice_id': pol.direccio_pagament.id,
-            'partner_id': pol.titular.id,
-            'account_id': 3,
-            'polissa_id': polissa_id,
-            'tarifa_acces_id': pol.tarifa.id,
-            'cups_id': pol.cups.id,
-            'factura_id': 3,
-            'llista_preu': pol.llista_preu.id,
-        })
-
-        atc_id = atc_obj.create_ATC_R1_009_from_polissa_via_wizard(
-            self.cursor, self.uid, polissa_id, {}
-        )
-
-        atc = atc_obj.browse(self.cursor, self.uid, atc_id)
-
-        tag_id = self.get_object_reference('som_autoreclama', 'tag_som_autoreclama_009_GEGC')[1]
-
-        self.assertEqual(atc.name, u"AUTOCAC 009")
-        self.assertEqual(atc.polissa_id.id, polissa_id)
-        self.assertEqual(atc.tag.id, tag_id)
-
-    def test_create_ATC_R1_009_from_polissa_via_wizard__tag_one_non_20TD(self):
-        atc_obj = self.get_model("giscedata.atc")
-        pol_obj = self.get_model("giscedata.polissa")
-        tar_obj = self.get_model("giscedata.polissa.tarifa")
-        f1_obj = self.get_model("giscedata.facturacio.importacio.linia")
-        ff_obj = self.get_model("giscedata.facturacio.importacio.linia.factura")
-
-        _, polissa_id = self.get_object_reference(
-            "giscedata_polissa", "polissa_0004"
-        )
-
-        tar_ids = tar_obj.search(self.cursor, self.uid, [("name", "!=", "2.0TD")])
-        pol_obj.write(self.cursor, self.uid, polissa_id, {
-            'mode_facturacio': 'atr',
-            'tarifa': tar_ids[0],
-        })
-
-        pol = pol_obj.browse(self.cursor, self.uid, polissa_id)
-        f1_ids = f1_obj.search(self.cursor, self.uid, [("cups_id", "in", [pol.cups.id])])
-        self.assertGreater(len(f1_ids), 0)
-        f1_id = f1_ids[0]
-
-        f1 = f1_obj.browse(self.cursor, self.uid, f1_id)
-        self.assertEqual(f1.cups_id.name, pol.cups.name)
-
-        f1_obj.write(self.cursor, self.uid, f1_id, {
-            "invoice_number_text": "1234567890ABCD",
-            "type_factura": "R",
-            "polissa_id": pol.id,
-            "fecha_factura_desde": "2015-01-01",
-            "fecha_factura_hasta": "2015-01-31",
-            "import_phase": 50,
-        })
-        ff_obj.create(self.cursor, self.uid, {
-            'linia_id': f1_id,
-            'tipo_factura': '04',
-            'importacio_id': 1,
-            'address_invoice_id': pol.direccio_pagament.id,
-            'partner_id': pol.titular.id,
-            'account_id': 3,
-            'polissa_id': polissa_id,
-            'tarifa_acces_id': pol.tarifa.id,
-            'cups_id': pol.cups.id,
-            'factura_id': 3,
-            'llista_preu': pol.llista_preu.id,
-        })
-
-        atc_id = atc_obj.create_ATC_R1_009_from_polissa_via_wizard(
-            self.cursor, self.uid, polissa_id, {}
-        )
-
-        atc = atc_obj.browse(self.cursor, self.uid, atc_id)
-
-        tag_id = self.get_object_reference('som_autoreclama', 'tag_som_autoreclama_009_GEGC')[1]
-
-        self.assertEqual(atc.name, u"AUTOCAC 009")
-        self.assertEqual(atc.polissa_id.id, polissa_id)
-        self.assertEqual(atc.tag.id, tag_id)
-
-    @unittest.skip(reason="WIP using mock")
-    @mock.patch.object(giscedata_polissa.GiscedataPolissa, "autoconsumo")
-    def __test_create_ATC_R1_009_from_polissa_via_wizard__tag_two_auto(self, mock_autoconsumo):
-        atc_obj = self.get_model("giscedata.atc")
-        pol_obj = self.get_model("giscedata.polissa")
-        tar_obj = self.get_model("giscedata.polissa.tarifa")
-        f1_obj = self.get_model("giscedata.facturacio.importacio.linia")
-        ff_obj = self.get_model("giscedata.facturacio.importacio.linia.factura")
-
-        mock_autoconsumo.return_value = 42
-        _, polissa_id = self.get_object_reference(
-            "giscedata_polissa", "polissa_0004"
-        )
-
-        tar_ids = tar_obj.search(self.cursor, self.uid, [("name", "=", "2.0TD")])
-        pol_obj.write(self.cursor, self.uid, polissa_id, {
-            'mode_facturacio': 'atr',
-            'tarifa': tar_ids[0],
-        })
-
-        pol = pol_obj.browse(self.cursor, self.uid, polissa_id)
-        f1_ids = f1_obj.search(self.cursor, self.uid, [("cups_id", "in", [pol.cups.id])])
-        self.assertGreater(len(f1_ids), 0)
-        f1_id = f1_ids[0]
-
-        f1 = f1_obj.browse(self.cursor, self.uid, f1_id)
-        self.assertEqual(f1.cups_id.name, pol.cups.name)
-
-        f1_obj.write(self.cursor, self.uid, f1_id, {
-            "invoice_number_text": "1234567890ABCD",
-            "type_factura": "R",
-            "polissa_id": pol.id,
-            "fecha_factura_desde": "2015-01-01",
-            "fecha_factura_hasta": "2015-01-31",
-            "import_phase": 50,
-        })
-        ff_obj.create(self.cursor, self.uid, {
-            'linia_id': f1_id,
-            'tipo_factura': '04',
-            'importacio_id': 1,
-            'address_invoice_id': pol.direccio_pagament.id,
-            'partner_id': pol.titular.id,
-            'account_id': 3,
-            'polissa_id': polissa_id,
-            'tarifa_acces_id': pol.tarifa.id,
-            'cups_id': pol.cups.id,
-            'factura_id': 3,
-            'llista_preu': pol.llista_preu.id,
-        })
-
-        atc_id = atc_obj.create_ATC_R1_009_from_polissa_via_wizard(
-            self.cursor, self.uid, polissa_id, {}
-        )
-
-        atc = atc_obj.browse(self.cursor, self.uid, atc_id)
-
-        tag_id = self.get_object_reference('som_autoreclama', 'tag_som_autoreclama_009_GEA')[1]
-
-        self.assertEqual(atc.name, u"AUTOCAC 009")
-        self.assertEqual(atc.polissa_id.id, polissa_id)
-        self.assertEqual(atc.tag.id, tag_id)
-
-    def test_create_ATC_R1_009_from_polissa_via_wizard__tag_tree_2_0TD_periods_auto(self):
-        atc_obj = self.get_model("giscedata.atc")
-        pol_obj = self.get_model("giscedata.polissa")
-        tar_obj = self.get_model("giscedata.polissa.tarifa")
-        f1_obj = self.get_model("giscedata.facturacio.importacio.linia")
-        ff_obj = self.get_model("giscedata.facturacio.importacio.linia.factura")
-
-        _, polissa_id = self.get_object_reference(
-            "giscedata_polissa", "polissa_0004"
-        )
-
-        tar_ids = tar_obj.search(self.cursor, self.uid, [("name", "=", "2.0TD")])
-        pol_obj.write(self.cursor, self.uid, polissa_id, {
-            'mode_facturacio': 'atr',
-            'tarifa': tar_ids[0],
-        })
-
-        pol = pol_obj.browse(self.cursor, self.uid, polissa_id)
-        f1_ids = f1_obj.search(self.cursor, self.uid, [("cups_id", "in", [pol.cups.id])])
-        self.assertGreater(len(f1_ids), 0)
-        f1_id = f1_ids[0]
-
-        f1 = f1_obj.browse(self.cursor, self.uid, f1_id)
-        self.assertEqual(f1.cups_id.name, pol.cups.name)
-
-        f1_obj.write(self.cursor, self.uid, f1_id, {
-            "invoice_number_text": "1234567890ABCD",
-            "type_factura": "R",
-            "polissa_id": pol.id,
-            "fecha_factura_desde": "2015-01-01",
-            "fecha_factura_hasta": "2015-01-31",
-            "import_phase": 50,
-        })
-        ff_obj.create(self.cursor, self.uid, {
-            'linia_id': f1_id,
-            'tipo_factura': '04',
-            'importacio_id': 1,
-            'address_invoice_id': pol.direccio_pagament.id,
-            'partner_id': pol.titular.id,
-            'account_id': 3,
-            'polissa_id': polissa_id,
-            'tarifa_acces_id': pol.tarifa.id,
-            'cups_id': pol.cups.id,
-            'factura_id': 3,
-            'llista_preu': pol.llista_preu.id,
-        })
-
-        atc_id = atc_obj.create_ATC_R1_009_from_polissa_via_wizard(
-            self.cursor, self.uid, polissa_id, {}
-        )
-
-        atc = atc_obj.browse(self.cursor, self.uid, atc_id)
-
-        tag_id = self.get_object_reference('som_autoreclama', 'tag_som_autoreclama_009_GET')[1]
-
-        self.assertEqual(atc.name, u"AUTOCAC 009")
-        self.assertEqual(atc.polissa_id.id, polissa_id)
-        self.assertEqual(atc.tag.id, tag_id)
-
-    # TODO: test NF readings
-"""
